Creating the webserver (servemux/router) and a 'home' handler ('controller')
The TCP network address should be in the format 'host:port', we can omit the host and will listen to our pc's available network interfaces. Sometimes, instead of a number people use ':http' or ':http-alt', that is a named port that Go will attempt to look up from /etc/services file when starting the server.