More routes, handlers and 404 catch
Servemux supports two patterns: fixed paths and subtree paths. Fixed paths like snippet/view, need to match exactly on the brower to use the snippetView handler. If not, such as the '/' route, it's a subtree, works like a catch-all. We can put a defense on the home handler. Request URL paths are automatically sanitized